qcacld-3.0: Sync RTPM get and put calls for STC client
Currently, when bursty flows are classified, STC prevents
RTPM suspend by calling RTPM_get and when the flow is
no more present, STC will allow RTPM suspend by calling
RTPM_put. In between if WLAN is turned off, during
de-register of STC module, RTPM_put is called to keep
get and put in sync when wlan is being deregistered.
Only bursty traffic should call runtime PM APIs. Thus,
use a flag to detect if there is a change in such flows
before acting on runtime PM APIs.

CRs-Fixed: 4134610

+++ components/dp/core/inc/wlan_dp_stc.h
@@ -491,7 +491,7 @@ struct wlan_dp_stc {
 	enum wlan_dp_stc_periodic_work_state periodic_work_state;
 	qdf_hrtimer_data_t flow_sampling_timer;
 	enum wlan_dp_stc_timer_state sample_timer_state;
-	uint32_t rtpm_control_flow_cnt;
+	uint16_t rtpm_control_flow_cnt;
 	bool rtpm_control;
 	struct wlan_dp_stc_peer_traffic_context peer_tc[DP_STC_MAX_PEERS];
 	struct wlan_dp_stc_sampling_table *sampling_flow_table;
